backend/app/main.py
# backend/app/main.py
from dotenv import load_dotenv
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

from app.api.health import router as health_router
from app.api.v1.auth import router as auth_router
from app.api.v1.findings import router as findings_router

load_dotenv()

app = FastAPI(title="Overmnd")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# mount routers
app.include_router(health_router, prefix="/api/v1")
app.include_router(auth_router, prefix="/api/v1")
app.include_router(findings_router, prefix="/api/v1")


# optional kube-style probe
@app.get("/healthz")
def healthz():
    return {"ok": True}
