backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.43
psycopg2-binary==2.9.9
pydantic==2.9.1
python-dotenv==1.0.1
email-validator==2.2.0
